Organize components directory
