+++ components/explore_modules/clickable_module_list/module_cards/card_structure.py
