+++ components/explore_modules/clickable_module_list/module_cards/modal_card_details.py
