+++ components/explore_modules/left_hand_nav_bar/search_panel.py
