+++ components/explore_pathways/pathway_card_details.py
